+++ js/google-auth.js
@@ -1,0 +1,22 @@
+//https://developers.google.com/identity/sign-in/web/backend-auth
+const CLIENT_ID = '504818718989-cpm7gtj8bjncqbjl3gp65tpdhj4uhman.apps.googleusercontent.com';
+const {OAuth2Client} = require('google-auth-library');
+const client = new OAuth2Client(CLIENT_ID);
+async function verify(token) {
+  const ticket = await client.verifyIdToken({
+      idToken: token,
+      audience: CLIENT_ID,  // Specify the CLIENT_ID of the app that accesses the backend
+      // Or, if multiple clients access the backend:
+      //[CLIENT_ID_1, CLIENT_ID_2, CLIENT_ID_3]
+  });
+  const payload = ticket.getPayload();
+  const userid = payload['sub'];
+  // If request specified a G Suite domain:
+  //const domain = payload['hd'];
+  return userid;
+}
+// verify().catch(console.error);
+
+module.exports = {
+    verify,
+}
